requirements.txt
flask
requests
torch
timm
numpy
opencv-python-headless
pillow
pandas
python-dotenv
sentence-transformers
faiss-cpu
prometheus-client
gunicorn
